update network check for single instance
As reported in #98, it is possible to abuse attributes in a manner that
the netowkr check will contain multiple instances.
This change will raise the problem to the user as a WARN, but still
enable them to render an incorrect file, as we don't know which
attribute invocation they would want.

A simpler workaround is to use override_attributes where necessary, as
the default_attributes may be getting appended to the exisiting default
array, as definced in the attributes/network.rb file.

Fixes #98